requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
Faker==4.1.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
importlib-resources==2.0.1
ipdb==0.13.3
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.6.2
prompt-toolkit==3.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.6.1
pypi==2.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.1.0
zipp==3.1.0
